requirements-ci.txt
numpy
scipy
sympy
tensorflow
tensorboard
networkx
quantum-blackbird
python-dateutil
thewalrus>=0.17.0
toml
appdirs
numba
requests
urllib3
plotly
